Register "Extra->Mixer"; break ExtraMenus -> MeterToolBar
