+++ src/toolbars/MeterToolBar.cpp
@@ -554,3 +554,77 @@ AttachedToolBarMenuItem sAttachment2{
 //   { PlayMeterBarID, RecordMeterBarID }
 //};
 }
+
+// Now define other related menu items
+#include "../commands/CommandContext.h"
+
+namespace {
+void OnOutputGain(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, true );
+   tb.ShowOutputGainDialog();
+}
+
+void OnOutputGainInc(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, true );
+   tb.AdjustOutputGain(1);
+}
+
+void OnOutputGainDec(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, true );
+   tb.AdjustOutputGain(-1);
+}
+
+void OnInputGain(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, false );
+   tb.ShowInputGainDialog();
+}
+
+void OnInputGainInc(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, false );
+   tb.AdjustInputGain(1);
+}
+
+void OnInputGainDec(const CommandContext &context)
+{
+   auto &project = context.project;
+   auto &tb = MeterToolBar::Get( project, false );
+   tb.AdjustInputGain(-1);
+}
+   
+using namespace MenuTable;
+BaseItemSharedPtr ExtraMixerMenu()
+{
+   static BaseItemSharedPtr menu{
+   Menu( wxT("Mixer"), XXO("Mi&xer"),
+      Command( wxT("OutputGain"), XXO("Ad&just Playback Volume..."),
+         OnOutputGain, AlwaysEnabledFlag ),
+      Command( wxT("OutputGainInc"), XXO("&Increase Playback Volume"),
+         OnOutputGainInc, AlwaysEnabledFlag ),
+      Command( wxT("OutputGainDec"), XXO("&Decrease Playback Volume"),
+         OnOutputGainDec, AlwaysEnabledFlag ),
+      Command( wxT("InputGain"), XXO("Adj&ust Recording Volume..."),
+         OnInputGain, AlwaysEnabledFlag ),
+      Command( wxT("InputGainInc"), XXO("I&ncrease Recording Volume"),
+         OnInputGainInc, AlwaysEnabledFlag ),
+      Command( wxT("InputGainDec"), XXO("D&ecrease Recording Volume"),
+         OnInputGainDec, AlwaysEnabledFlag )
+   ) };
+   return menu;
+}
+
+AttachedItem sAttachment4{
+   wxT("Optional/Extra/Part1"),
+   Shared( ExtraMixerMenu() )
+};
+
+}
